divided fast send function
